215_kamiooya/C1_cursor/1b_Cursorマニュアル/requirements_gmail.txt
# Gmail → やり取り連携スクリプト用
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
PyYAML>=6.0

# Notion タスク登録スクリプト用
requests>=2.28.0
